feat: replace modals with dedicated pages for readme and download
Replaced the popup modals on the home page for reading the project README and downloading the latest APK with dedicated routes (`/readme` and `/download`).

This improves visibility on all screen sizes and provides a more integrated experience within the website, while maintaining the dynamic fetching of data directly from GitHub.
